Release 0.5.0
- Added command as argument syntax with better
  performance and ability to record & replay stderr

- Added --quiet option

- Changed targeted timers to remove ANSI codes

+++ lib/timmy/logger.rb
@@ -5,6 +5,10 @@ module Timmy
         @output_dir = File.expand_path(dir)
       end
 
+      def set_quiet(quiet)
+        @quiet = quiet
+      end
+
       def set_precision(precision)
         @precision = precision
       end
@@ -13,38 +17,61 @@ module Timmy
         @profile = profile
       end
 
-      def put_output(output)
-        duration = MasterTimer.get
-        formatted_duration = format_duration(duration)
+      def match_replay_header(line)
+        line.match(/^TIMMY-SESSION:v1:(?<s>\d+\.\d{9})$/)
+      end
 
-        puts feint(formatted_duration) + " " + output
+      def match_replay_line(line)
+        line.match(/^(?<s>\d+(\.\d+)?)(?<t>e)? (?<content>.*)/)
+      end
+
+      def put_output(output, error = false)
+        duration = MasterTimer.get
+
+        if @quiet
+          puts output
+        else
+          formatted_duration = format_duration(duration)
+          formatted_duration = red(formatted_duration) if error
+          puts "\e[0m" + feint(formatted_duration) + " " + output
+        end
+        $stdout.flush
 
         @output ||= ''
-        @output += sprintf("%.9f %s\n", duration, output)
+        @output += sprintf("%.9f%s %s\n", duration, error ? 'e' : '', output)
       end
 
       def put_eof
-        put_output(feint("EOF"))
+        put_output(feint("EOF")) unless @quiet
       end
 
       def put_timer(timer)
-        puts format_timer(timer)
+        do_put_timer(timer) unless @quiet
       end
 
       def finalize
-        suffix = "#{MasterTimer.start.to_i}+#{MasterTimer.get.to_i}"
-        filename = File.join(output_dir, "timmy-#{suffix}.log")
-        header = sprintf("TIMMY-SESSION:v1:%.9f\n", MasterTimer.start)
-
-        File.write(filename, header + @output)
-
-        puts feint("Log written to #{filename}")
-        puts
-
+        save
         put_profile if profile?
       end
 
       private
+
+      def save
+        suffix = "#{MasterTimer.start.to_i}+#{MasterTimer.get.to_i}"
+        filename = File.join(output_dir, "timmy-#{suffix}.log")
+
+        return if File.exists?(filename)
+
+        header = sprintf("TIMMY-SESSION:v1:%.9f\n", MasterTimer.start)
+        File.write(filename, header + @output)
+        puts feint("Log written to #{filename}")
+        puts
+      end
+
+      def do_put_timer(timer)
+        puts format_timer(timer)
+        $stdout.flush
+      end
 
       def put_profile
         slowest_timers = TargetedTimerManager
@@ -64,9 +91,14 @@ module Timmy
       end
 
       def format_timer(timer)
-        string = "#{bold(format_duration(timer.duration))} #{format_id(timer.definition.id)}"
-        string += " (#{timer.group})" if timer.group
-        string += ": #{green(timer.label)}" if timer.label
+        string = (
+          bold(format_duration(timer.duration)) +
+          " " +
+          bold(green(format_id(timer.definition.id)))
+        )
+
+        string += " " + green("(" + timer.group + ")") if timer.group
+        string += " #{timer.label}" if timer.label
 
         string
       end
@@ -80,16 +112,20 @@ module Timmy
         sprintf(format, duration / 60, duration % 60)
       end
 
-      def green(string)
-        "\e[0m\e[32m#{string}\e[0m"
-      end
-
       def bold(string)
-        "\e[0m\e[1m#{string}\e[0m"
+        "\e[1m#{string}\e[0m"
       end
 
       def feint(string)
-        "\e[0m\e[2m#{string}\e[0m"
+        "\e[2m#{string}\e[0m"
+      end
+
+      def green(string)
+        "\e[32m#{string}\e[0m"
+      end
+
+      def red(string)
+        "\e[31m#{string}\e[0m"
       end
 
       def output_dir
